requirements.txt
boto3==1.40.49
botocore==1.40.49
pandas==2.3.3
ipykernel==6.30.1
tqdm==4.67.1
fsspec==2025.9.0
huggingface_hub==0.35.3
gtts==2.5.4
mutagen==1.47.0
fastparquet==2024.11.0
transformers==4.57.0
datasets==4.2.0
bs4==0.0.2
dotenv==0.9.9
matplotlib==3.10.7
pillow==11.3.0
piper-tts==1.3.0
nbformat==5.10.4
librosa==0.11.0
pydub==0.25.1
soundfile==0.12.1
audioread==3.0.1
ffmpeg-python==0.2.0
numpy>=2.0.0
scipy==1.13.1
chromadb==1.2.0
sentence-transformers==5.1.1
opencv-python==4.12.0.88
streamlit==1.50.0
streamlit_audiorecorder==0.0.6
accelerate==1.11.0
google-genai==1.46.0
open-clip-torch==2.31.0
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
wav2clip
seaborn
st-pages==1.0.1
torchcodec==0.8.1